Deprecate torchsampler.__about__
It's better to keep all metadata in setup.py and use importlib.metadata.
Otherwise, you need to 'import __about__' during the build process
means importlib-metadata needs to be installed during the build
if building on python 3.6 or python 3.7, and potentially running an
unbounded amount of other code during the build.

It's easier to just sidestep the issue.